refactor scope manager
Summary:
The old scope manager was a fork of `eslint-scope`, which has a pretty ugly codebase - long, sprawling files and was originally written without types and was not written for type analysis.

This approach could work, but it's difficult to get completely right because the type-only logic was intermingled with the value logic. This lead to some bugs where a type-location `Identifier`s could incorrectly create and reference values.
Additionally, `eslint-scope` has a few bugs that have never been fixed due to lack of motivation amongst other things.

`typescript-eslint/scope-manager` is a fork of `eslint-scope`, but it was rewritten from the ground-up to support type scope analysis. It's designed to compartmentalise the type-scope analysis into a separate visitor, which means it's impossible to create incorrect references. It also does not have the same open bugs.

This diff:
- forks `typescript-eslint/scope-manager`, and switches out the TS-specific code for the old flow-specific code. This brings along the following changes:
  - codebase is split up into a number of modules rather than a few monolithic files.
  - remove some functionality that has been deprecated by ESLint and isn't used in any modern rules (like `isArgumentsMaterialized`)
  - remove logic based on `ecmaVersion` - we don't need to worry about this as flow treats code as the latest version it supports.
  - remove the type scope from a few places so the scope tree is more consistent with the TS scope tree to aid with OSS tooling.
  - change the `isTypeReference`/`isValueReference` methods to be getters - which is in-line with the TS-scope manager to aid with OSS tooling.
  - [NEW] updates class properties to include nest a `class-field-initializer` scope, inline with the spec.
  - [BUG] fixes a bug that caused references to class supers being placed in the upper scope instead of within the class scope.
  - [BUG] fixes a bug that meant that globally scoped variables would not have their references closed by the scope manager.
  - [NEW] correctly define type/value imports so that they can only be referenced within type/value locations respectively.
- flowify and fix tests for the above changes.
  - also updated jest config so that the tests can be run against `src` rather than `dist`

In the next diff I shall fork the testing tooling as well and add comprehensive tests.

Reviewed By: pieterv

Differential Revision: D33290882

fbshipit-source-id: 60e2f2c6945625a7bd361ac402abf336bbdfc7f2