Separate stats submission from CollectionStats destructor
Summary:
Collection stats are currently submitted implicitly, in the destructor
of the `CollectionStats` object. Having this side effect in the
destructor is hard to reason about, since we cannot guarantee that the
stats are submitted at well defined points in the program. For
instance, during an OOM, we may be in the middle of a YG collection. In
that case, the YG stats would be incomplete, but we'd still end up
attempting to submit them via the analytics callback.

In upcoming diffs, we will also add the ability to destruct `HadesGC`
without completing pending OG collections, so we want the ability to
discard any pending stats.

Reviewed By: kodafb

Differential Revision: D33390634

fbshipit-source-id: df1757e3b9e60dc9141af38b58ea8b00c4591771