Do not collect OG during YG collections
Summary:
It is unsafe to perform OG collection work during a YG collection
because WeakMap marking attempts to access objects on the heap, which
may have been forwarded. This change stops blocking on OG collections
when we're under memory pressure or during an OOM.

This should be ok since:
1. The behaviour is already broken, and should not be used.
2. This can only happen in the very rare case where we cannot allocate
any more segments for the heap, at which point an extra OG collection
may not be of much use.

Reviewed By: kodafb

Differential Revision: D33390633

fbshipit-source-id: e73bd10ab24f7b6080a7584fd29ceb268840fd3c