fix handling of imports
Summary:
All imports are treated as "type" definitions regardless of whether or not they have a "type" kind. You can import classes and enums as values and use them in type locations. Additionally namespace imports can be used in a type location (eg `React.MixedElement`).

Because we don't have cross-file information, we cannot statically determine if an imported thing is a "type referencible" thing, so we have to just treat all imports as possible types.

This diff marks all import bindings as type definitions.

Reviewed By: evanyeung

Differential Revision: D33574248

fbshipit-source-id: b6dc10dc2afb8ad01e0dbfbe40d0dab25b0ec297