add generated predicate functions
Summary:
Predicate functions are shortcuts for doing various checks.
Some people have asked for these to be exposed because they prefer them to manual checks.

This diff codegens a predicate function for each node and for a large number of tokens.
It also adds a few "custom" predicate functions which also shorthand some common checks

Reviewed By: pieterv

Differential Revision: D33933801

fbshipit-source-id: 01c0b4c8d52bacebbcc9383af4754d8e94d891cd