Add replay tests for synth trace
Summary:
We currently don't have a convenient way to write unit tests for synth
traces that exercise the entire pipeline from recording to replay. This
adds a new test fixture that makes it easier to write such tests.

This is particularly useful when adding new things to jsi/synth traces,
since we don't have existing synth traces that will exercise the new
functionality.

Reviewed By: kodafb

Differential Revision: D34062870

fbshipit-source-id: c9484ae45dd973866c619948830063625bd7cd79