Add SemContext::dump
Summary:
Add a function to dump information about all the information in a
given `SemContext`.

It prints directly to stdout for now because the intent is only to use it
for test/debug situations, but it can be swapped to use a real
`BufWriter` if needed in the future. It requires a `GCLock`, so we can't
use methods such as `impl Display` anyway.

Reviewed By: tmikov

Differential Revision: D33557374

fbshipit-source-id: d55da52bb1574515e49757e168f436b0d09029c7