Add getCellKind for cell classes
Summary:
Add the ability to get the `CellKind` for a given cell class. This
already exists for a small number of cells (the ones that are
templates), this diff expands it to everything else.

This is used in a subsequent diff to retrieve the `CellKind` for
classes in the `GC::makeA` template.

Reviewed By: kodafb

Differential Revision: D33251143

fbshipit-source-id: 7236eac5e23832c59faf2052913cf7d49e26eda4