Simplify drain rate calculation
Summary:
The current calculation needs to round up in 3 places in order to avoid
division by zero at the end. We can instead rewrite the formula so only
`bytesToFill` needs to be non-zero.

This can have slightly different results in the zero case, since the
max drain rate was previously `(preAllocated - markedBytes)`, but is
now `(preAllocated - markedBytes) * HeapSegment::maxSize`. That doesn't
really matter since we're limited by `(preAllocated - markedBytes)`
anyway. However, in the common case where `ygCollectionsUntilFull > 1`,
the range of values should be roughly the same as the ones we get from
the current formula.

Reviewed By: kodafb

Differential Revision: D34160580

fbshipit-source-id: 44e8fbfdf5aa7c0db8f028aae7d9d46c21fb3504