Include external memory in target heap size
Summary:
Hades starts collections based on the occupied fraction of the heap,
which is given by the total allocated bytes divided by the total size
of the heap. The problem is that in the denominator, it is hard to know
what we should consider as the total size of external memory, so we
just put the external memory bytes in both the numerator and
denominator, effectively always treating the external memory as having
an occupancy ratio of 1.

This interacts badly with the mechanism we have for sizing the managed
portion of the heap. That mechanism grows the target size of the
managed portion in response to big increases in external memory usage,
in order to avoid ending up in a state where we are constantly
collecting. The core problem here is that the external memory size and
the target size are tracked as separate metrics, but they can influence
each other in ways that create bad feedback loops. Here is a
hypothetical example to illustrate, given an occupancy ratio of 50% and
a collection threshold of 90%:

1. End of first collection, allocated bytes = 5MB, external bytes =
200MB. New target size for the managed portion is 210MB, leading to a
post collection occupancy ratio of 50%.
2. For the start of the next collection to happen at 90%, if there is
no substantial increase in allocated bytes, we need to solve for (5 +
N) / (210 + N) = 0.9. That gives N = 1.84GB.

This diff incorporates external memory into the target heap size, so
that we have a fixed denominator based on the total live data volume in
the heap at the end of the last collection.

This is a perf regression and a memory win due to an increase in the
number of collections from the new heuristics. I think the heuristics
are reasonable, it's just a matter of tuning the occupancy ratio and
the collection threshold to move around on the memory/perf tradeoff.

Reviewed By: kodafb

Differential Revision: D33910828

fbshipit-source-id: bb3ea7f0bf119870c846fb1d406d68cad3b2967c