Use an opaque list structure in SemContext
Summary:
To allow the borrow checker to see that we're borrowing separate fields
using the `decls` and `scopes` functions, expose the fields directly.
We don't show the `Vec`s as a raw type to avoid having to directly index
them, instead creating an opaque list type which allows us to index
using the various existing `Id` types.

Reviewed By: tmikov

Differential Revision: D34011961

fbshipit-source-id: 91755464c6bdf432d04f94599205a31eb61f860c