Detect local eval
Summary:
Set a flag in the function that uses local `eval` as well as all its
ancestors, so we know that their variables may be modified by the `eval`.

In non-strict mode, `eval` may introduce new local vars, so all
descendent functions must be checked and if they use variables
declared in ancestors of the `eval`, they must be marked Unresolvable.

Add a `can_rename` flag which indicates whether variables themselves
can be renamed, which they can't in the presence of local `eval` in
their descendents.

Reviewed By: tmikov

Differential Revision: D33898745

fbshipit-source-id: 98a879a55588f622d53956003f700753a0ed6363