Implement promote_scoped_func_decls
Summary:
Add the function to promote locally scoped function declarations
to function scope.

This requires some minor changes to the API of the `DeclCollector`,
which has a function to add an entry to the function scope after the fact,
but requires that no `Rc`s are outstanding when that happens.

Reviewed By: tmikov

Differential Revision: D33645181

fbshipit-source-id: 2fe93550294c4ecc1a4d09f7934d04f784fd9f02