Correctly set prototypes that should be plain objects
Summary:
According to the spec, `Map`, `Set`, and `Error` prototypes are just
plain objects, not instances of their respective types. (the behaviour
for Error changed between ES5 and ES6)

https://tc39.es/ecma262/#sec-properties-of-the-error-prototype-object
https://tc39.es/ecma262/#sec-properties-of-the-set-prototype-object
https://tc39.es/ecma262/#sec-properties-of-the-map-prototype-object

Note that we are already doing the right thing for
`typedArrayBasePrototype` but the code is misleading as written.
`TypedArrayBase` does not have a `create` method, so it just invokes
create on the base class (i.e. `JSObject::create`).

Reviewed By: kodafb

Differential Revision: D34411062

fbshipit-source-id: 29856e0d2fc024b68295b2c4733442aed4433106