Set CellKind and size in GC
Summary:
We set the kind and size of a cell in the GCCell constructor, by
passing in a VTable. By using `getCellKind` and the allocation size, we
can now do this entirely from the GC. This has a few advantages:
1. This gives us a new invariant that cell headers may only be written
by the GC, which is important because they are synchronised by
`gcMutex_`. A future diff could now shrink the critical section in the
OG alloc path to only touch the header, instead of holding the mutex
for the whole constructor.
2. This is more efficient than looking up the kind and size in the
VTable, especially if the compiler is not able to optimise it.
3. With the next diff, this will require us to only pass the size into
the constructor once for variable sized cells.

Reviewed By: kodafb

Differential Revision: D33251139

fbshipit-source-id: f866841e9dc4bffdd97d8c854eff4ed233127663