Restore anonymous internal properties
Summary:
We previously separated "anonymous" internal properties from "named"
internal properties. The former described properties that were only
used to reserve hidden class slots and the latter could be used to
store VM state on objects.

These were removed in D25373892, and it wasn't really a issue because
named internal properties were not actually being used. However, now
that we are using them for structured errors and soon for Intl, this
distinction is important.

In particular, we have many checks that currently use the total number
of internal properties when they should in fact be checking only the
number of anonymous internal properties. The way the code is
structured, we could end up creating an object that uses a named
internal property to reserve a hidden class slot that is in fact being
used for something else (either overlap slots or as an "additional
property"). This could mean for example, that an object appears to have
a CapturedError internal property even though it actually just uses
some extra internal slots.

Reviewed By: jpporto

Differential Revision: D35203678

fbshipit-source-id: 89bf03b2043001c6836625de16c35e9f8164b0d0