Fix out of bounds mutations
Summary:
So the `insertBeforeStatement` mutation was always off by one, it contained an unnecessary `- 1` when specifying the array index to splice. Interestingly this was not discovered via the `InsertBeforeStatement` unit tests due to them only containing one item this caused the insert index to be `-1` which when used with splice would go backwards from the end, so inserting at position `0`, the correct spot! :D

To solve this:
- Fix the bug.
- Update `arrayUtils.js` to bounds checks so they types of issues are found.
- Write unit tests to cover varied array sizes and positions.

Reviewed By: kassens

Differential Revision: D35230898

fbshipit-source-id: b368bddf209ef2621ccf6ab3c0d06541f62fe0b3