Generalise handling failure to set internal properties in structured errors
Summary:
1. Move CallSite to using `defineNewOwnProperty` instead of
`defineOwnProperty`, since the object was just created. This brings it
in line with the usage in Intl. It also removes the exception checks
after the assert that there was no exception.
2. Consolidates error handling in `defineOwnProperty` in
`captureStackTrace`. When an object is not extensible, it returns
false, and we can then use that to return an error to the user.

Reviewed By: jpporto

Differential Revision: D35170458

fbshipit-source-id: 83398362c38b4c64d5c28fc95027dab7eb9c5852