Enable ICF and LTO (#720)
Summary:
Turn on ICF for the build whenever it is available. This requires us to
link with LLD, which is only the default starting in NDK r23, so
manually specify LLD for the Android build.

Also turn on LTO for the CircleCI build, so we generate highly
optimised binaries for distribution.

Pull Request resolved: https://github.com/facebook/hermes/pull/720

Differential Revision: D35344254

fbshipit-source-id: e4e6c7fdbf738395054e3ddc3b00563ca6fbb73d