Use ASR for tag checking
Summary:
jpporto pointed out that it's possible to to make the HermesValue tag
checks cheaper by making them negative, and then using an arithmetic
right shift to get the tag. This means that instead of being a 16 bit
(or 17 bit for ETag) positive value, the tag becomes a small negative
value instead. This works because most of the top bits of the tag were
already 1 anyway.

On most platforms, loading a 1 byte signed immediate is going to be
cheaper than loading a 3 byte unsigned immediate, so tag comparisons
are faster.

Reviewed By: jpporto

Differential Revision: D35103174

fbshipit-source-id: 34739e1ceb20684a5fa9c4888c9ecd8404bfda15