Add HeapRuntime
Summary:
Add the ability to instantiate a runtime in the first segment on the
heap. `HeapRuntime` manages the lifetime of the `Runtime`, and ensures
that it is destructed properly.

For now, leave instantiation of the `Runtime` itself to the user of
`HeapRuntime`, to avoid needing to deal with making the constructor
accessible and making the arguments line up between the constructors of
`DummyRuntime` and `Runtime`.

Instead of mapping the full 4GB, allocate something close to the
specified max heap size, so that we don't sit on too many large
allocations (which can be a problem, particularly during tests). Since
honouring the max heap size is the responsibility of the GC, add a few
segments of buffer to the max size for GC internal behaviour (e.g.
compaction in Hades).

Reviewed By: kodafb

Differential Revision: D32422492

fbshipit-source-id: 0290425a8deb699b7a768da3bbc2c551287cd871