Fix trailing elision in variable length arrays (#729)
Summary:
We got a report that Hermes does not correctly set the length of arrays
that contain spread operators and trailing empty values. This is
happening because the existing code for trailing empty values does not
handle the variable length case. This diff makes sure that trailing
empty values are handled in all cases.

Closes #729

Reviewed By: jpporto

Differential Revision: D35549692

fbshipit-source-id: 2dd47be26a2df037050492c3f728cdf6fe8f46b8