Allow Hermes version to be set externally
Summary:
Allow the user building Hermes to either override or drop the release
version from Hermes. This allows users who are building Hermes from
source at commits that don't line up with Hermes versions to display a
more accurate version.

With this change, passing in an empty version will cause the version to
be dropped, passing in nothing will use the default version (which for
now is still going to be the last Hermes release), and passing in any
other string will cause that to be used as the version number displayed
by Hermes.

Reviewed By: jpporto

Differential Revision: D35710651

fbshipit-source-id: 4470648df55a6a6388b1b3b053797523fc6aba23