Visit block statements in Program nodes
Summary:
Some nodes with BlockStatement children handle the extra scope directly,
but Program is not one of those nodes, as it can have multiple statements,
e.g.

This was resulting in `let` inside a block scope:
```
{ let x; }
```
being resolved as an undeclared global instead of detecting
a declaration as it is supposed to.

Reviewed By: tmikov

Differential Revision: D34253708

fbshipit-source-id: 00caee25525648c92e94795f771dfe05f5298c80