Back WeakRef with WeakRoot
Summary:
Swap out the underlying storage in WeakRefSlot with a WeakRoot instead of a HermesValue. This will also open up code reuse with marking and keeping track of WeakRefs and WeakRoots.

Note that the value stored in `WeakRefSlot` actually has to be a union of `WeakRoot` and `WeakRefSlot *`. This is because a `WeakRefSlot` is treated as a node in a linked list of other `WeakRefSlot`s when it is freed.

Reviewed By: neildhar

Differential Revision: D35225408

fbshipit-source-id: 25d71c4f470f2c475c2b2c243ec14559dd7d7dd4