Add PointerBase/Runtime parameters to WeakRef(Slot) methods
Summary:
This is solely a preparation diff to make the list of files touched in upcoming diffs more manageable- it doesn't make sense as a standalone one.

In the upcoming diffs, most WeakRef and WeakRefSlot operations will need access to a pointer base to fully read the underlying CompressedPointer. The read barrier variants also will need a `GC` still, so it's easier to pass a `Runtime` which can act as both a `PointerBase` and has access to a `GC`.

Reviewed By: neildhar

Differential Revision: D35692840

fbshipit-source-id: 5230526978c8e4524fb04f2dec461e4a1a5e2dbb