Use AtomU16 for string literals
Summary:
Add more calls to allow storing string literals as `AtomU16`
in the atom table.

This removes all the malloced memory from the AST,
bringing down the time to free the AST substantially.

Tried to use generics to remove the substantial code duplication,
but couldn't find the right way to trait bounds this so it worked.

Reviewed By: tmikov

Differential Revision: D35153579

fbshipit-source-id: aa52b5bc36c336825b747c8b4ec3e14eb69ce556