Extract an append-only Deque for storage
Summary:
For separation of concerns and for easier reuse,
extract a `Deque` struct which can handle append-only operations
which are necessary for allocating new AST nodes.

The new struct is guaranteed to return references which will
never move, because none of the `Vec`s comprising the `Deque`
will ever reallocate.

The initial/maximum chunk sizes are large fixed numbers for now,
but can be made configurable with const generics if we want
in the future.

Intentionally omit the free list from `Deque` for simplicity
and force `Context` to continue to manage the context ID and free list.

Place the `Deque` in `juno_support` for easy use in other applications
such as in the IR.

Reviewed By: tmikov

Differential Revision: D35193543

fbshipit-source-id: 61f76348af1d54ccfc91207850ab283031009d00