Introduce HeapSize trait
Summary:
In order to make it easier to determine how much heap Juno is actually
using, introduce a `HeapSize` trait and a function to estimate the
heap size of a given data structure.

The trait is implemented for some simple structures, but estimating
size of `HashMap` isn't entirely accurate.

Implement `HeapSize` for `Context`.

Add an `--Xmem` flag to allow looking at heap size to test this.

Reviewed By: tmikov

Differential Revision: D35508865

fbshipit-source-id: cff9587dc5bde4fb393cc9e4b58654f33280cd7a