Detect flow pragma
Summary:
Certain use cases require the docblock to detect whether or not
to parse the file as Flow-typed.

Add a Juno dialect `flow-detect` which will use the docblock to detect
whether the file should be parsed as proper Flow or just unambiguous
Flow. Detection is done via calling the `hasFlowPragma` function
in the Hermes API.

Reviewed By: tmikov

Differential Revision: D35754811

fbshipit-source-id: f3e0eb84c3127620cae8a4bd3eb3cfb433509aa5