support ESTree spec for optional chaining
Summary:
This was a tricky one to implement! Unfortunately there was one case that I couldn't actually support due to limitations in the AST that the underlying hermes parser emits.
Otherwise - it was possible to fully support the ESTree optional chaining representation!
This was one of the major blockers for allowing flow codebases to upgrade to a modern ESLint version ([v7.5.0 was the first version which supported the ESTree spec](https://eslint.org/blog/2020/07/eslint-v7.5.0-released#optional-chaining-support)).

Reviewed By: pieterv

Differential Revision: D36622146

fbshipit-source-id: 3a3d7cd8e169e78dee65a8665778d3be61a0d3f7