Fix for empty files
Summary:
We need to account for empty `DataRef`s on both the start and end
of empty nodes.

First, we extend `DataRef` by 1 in the case that it's empty to avoid
failing to find an offset in the line.

Then to account for the `.pred()` call causing out of bounds reads
off the front of the line, we conditionalize it on whether the
range itself is empty. If the range is empty, we can just reuse
the start location.

Reviewed By: tmikov

Differential Revision: D36726860

fbshipit-source-id: b2da45454ef245e8c1ba328d407220f436c4a227