Hide dependencies of libhermes
Summary:
By default, all dependencies of a library in CMake are PUBLIC, which
means that anything depending on a library will also link against its
dependencies. In the case of libhermes for instance, if someone were to
include Hermes from source in their CMake project and take a dependency
on libhermes, their library would then also link against each of
libhermes' dependencies, which may be static libraries.

This is bad, since it depends on link order to resolve the symbols to
the libhermes dynamic library as opposed to the other static libraries.
It also means that if we accidentally fail to export something from
libhermes, it could still silently work but point to a different
symbol, which can cause incorrect behaviour at runtime.

This isn't a risk in most cases, because most users in open source
consume Hermes as a prebuilt library. But we should still fix the case
where Hermes is a CMake subproject.

Reviewed By: jpporto

Differential Revision: D36873153

fbshipit-source-id: 86ccf9f136d2b858467a3ab59754d98c966f599e