Add an explicit commit for virtual memory
Summary:
Add `vm_reserve` and `vm_commit` functions to OSCompat. These are
intended to be used to reserve large address spaces, when we are able
to explicitly commit memory before it is actually accessed. This makes
it significantly more likely that the OS will complete the
allocation.

This diff only implements the mechanism for OSCompatPosix, since we
don't currently use the contiguous heap on Windows. However, this API
should be straightforward to implement in terms of `VirtualAlloc` and
`VirtualFree` on Windows.

The implementation is based on guidance from the kernel documentation:
https://www.kernel.org/doc/html/v4.19/vm/overcommit-accounting.html

Specifically, if you create a mapping that is not writable, it does not
count against the overcommit accounting on Linux. This also uses the
`MAP_NORESERVE` flag, which is an additional hint to the kernel
(depending on the overcommit handling mode) that the availability of
backing storage should not be considered when fulfilling the
allocation.

Reviewed By: kodafb

Differential Revision: D36921482

fbshipit-source-id: 914ffd5b0f4b13ba90dde080961343db2af5cf2a