Directly access scopes in current function
Summary:
If a variable is being accessed in a scope in the same function, we
have access to that scope directly, so we don't need to traverse the
scope chain. If it is being accessed in a scope in a different
function, we can start the traversal from the function scope.

Reviewed By: tmikov

Differential Revision: D36926931

fbshipit-source-id: 010122cc3f5c8bb20e0d573fba82f0d2b86dc939