Independently collect ManagedValues
Summary:
We currently collect `ManagedValues` during root scanning in garbage
collection. However, the creation of `jsi::Value`s is not necessarily
correlated with the allocation of things on the heap. That is, we could
create a lot of `ManagedValues` without necessarily triggering a GC.

This can cause problems in applications that create `jsi::Value`s at a
high rate, since values that are no longer referenced will still be
kept around until the next GC.

To fix this, we can collect `ManagedValues` independently of GC cycles.
On creating a value, we can check whether it causes us to exceed some
target size, and if it does, we can free unused values.

The target size is computed by taking the number of values still in use
at the end of a collection, and dividing it by the occupancy ratio of
the heap. This ensures that the size overhead of maintaining
unreachable values is in line with the overall size overhead specified
for the heap.

Reviewed By: jpporto

Differential Revision: D36799075

fbshipit-source-id: a0b26962bcc991220fcb5088657401c966cd565f