Core IR: Value, Instruction, ValueList
Summary:
Implement the core IR, similar to LLVM IR. The main entities are Values
and Instructions tied together by the key data structure ValueList. An
Instruction contains a list of operands, each of which is a reference to
a Value. Each Value maintains a list of all of its users. The bulk of
this happens in a ValueList, which is the operand list in an
instruction *and* contains the links of the user list of a Value.

Reviewed By: neildhar

Differential Revision: D37448546

fbshipit-source-id: 7cc1448682116f0125188c24da7bc67c9798bac9