fix unnecessary newlines being added during printing
Summary:
There are cases where prettier inserts spaces in between detached nodes.
Mainly this impacted function arguments as we'd do things like:
```
[].map(foo(a, b) {})

// after replacing with an arrow became:

[].map(
  (
    a,

    b,
  ) => {},
)
```

This is bad for obvious reasons - because it is purely junk lines.
From extensive testing, I found that the issue ONLY occurs if there is a docblock.

My hypothesis is that when prettier lays out things like function parameters it looks to see if there is a comment on the line. If there is then it puts the param on the next line.
Though - I don't have a true root cause for this as I haven't properly dug into prettier's source code and it's a pain to do remote nodejs debugging from a devserver and I cannot be bothered checking out locally.

I tried a few fixes, but from my testing this fix (setting detached nodes to have range `[1, 1]`) works for cases of brand new detached nodes, and cloned detached nodes.

Reviewed By: pieterv

Differential Revision: D37499940

fbshipit-source-id: 6398186696692c047a6e4cb65a3901d57ef8745b