Always compact the last segment in the heap
Summary:
We currently allow compaction to select any segment in the OG. In
theory, this allows us to select the most fragmented segment for
compaction. In practice however, compaction primarily serves to shrink
rather than defragment the heap, since coalescing free blocks does a
good enough job. In fact, the current compaction heuristic just picks
the segment with the fewest allocated bytes.

Given this, it is much simpler to just compact the last segment in the
heap. This is an improvement for a few reasons:
1. Removing the last segment from the freelists is significantly
simpler.
2. It reduces the need for tracking the allocated bytes in each
segment, which allows us to remove that mechanism entirely later in the
stack.
3. The last segment is likely to be the emptiest anyway, so it lines up
with the current heuristic.

Reviewed By: kodafb

Differential Revision: D37445374

fbshipit-source-id: 2b4c6eaaff54b763be3ed6ec4d9882c996aff422