Stop tracking per-segment allocated bytes
Summary:
This diff removes the need to track the number of bytes allocated in
each segment, which simplifies the allocation path, and enables further
improvements to the allocator.

1. Removes the per segment allocated bytes counters
2. Starts tracking the compactee as part of the overall allocated bytes
and size counters for the OG, since we no longer know exactly how much
is allocated in the compactee when compaction starts.
3. Updates the finalisation pass for the compactee to also compute the
number of bytes allocated in the compactee, so that we can subtract it
from the OG's allocated bytes counter.

Reviewed By: kodafb

Differential Revision: D37445373

fbshipit-source-id: 2e1fcc355d839077edabf1a9d7639cb10ed8a02d