Fix underflow when calculating post-collection allocated bytes in Hades
Summary:
When calculating the post-collection allocated bytes, Hades takes the
pre-collection allocated bytes and subtracts the number of bytes swept.
This lets us calculate the number of bytes that survived the
collection, since objects allocated after the start of the collection
will not be collected.

However, those objects may be trimmed. For instance, arrays that are
longer than necessary may be shrunk by the GC. This means that in
extreme scenarios, we can appear to free more memory than was initially
allocated in the heap.

This breaks the invariant that the number of bytes swept must be
smaller than the number of bytes initially allocated, resulting in an
underflow when we try to calculate the post-collection allocated
bytes.

This diff fixes that by clamping the number of swept bytes to be no
larger than the number of pre-collection allocated bytes. It also
tracks the number of bytes trimmed in debug builds to ensure that any
discrepancy is accounted for by trimming.

Reviewed By: jpporto

Differential Revision: D37708886

fbshipit-source-id: a1233cd7a1f04c7019272e6987744ac0c62e8b60