Factor out helper for freeing internal data block
Summary:
Both the finaliser and `detach` methods of `JSArrayBuffer` duplicate
the steps needed to free the data block held by the buffer. Merge the
two implementations, using the version from the finaliser as a
reference.

This also fixes a current bug, where detaching does not call
`untrackNative` to tell the ID tracker that the allocation is being
freed.

Reviewed By: jpporto

Differential Revision: D37735990

fbshipit-source-id: 0b4d269c3f635789ed8dbfd852b73e172b2d9f55