Simplify state management of JSArrayBuffer
Summary:
`JSArrayBuffer`'s state currently has some undocumented invariants in
the presence of `detach`. Specifically, a detached buffer must have its
data and size set to zero.

This diff relaxes those invariants, so that the data and size only need
to be valid when the buffer is attached. This reduces the number of
places where we need to initialise the fields, which is useful as later
diffs in the stack add an additional field.

Reviewed By: jpporto

Differential Revision: D37735991

fbshipit-source-id: d147fbf1da6dd68bc09aa138d24d410e9f21cbd0