Refactor integer parsing and validation
Summary:
Split out the validation to check whether a given string is a valid
integer in the given radix. This lets us reuse the validation code for
checking that a `BigInt` is valid, instead of needing to call the
heavier weight `parseIntWithRadix` function, which performs some
additional operations on the resulting double.

Reviewed By: tmikov

Differential Revision: D37764391

fbshipit-source-id: c3b417b4a7a4ea8f5108ad741b8ad20e5729055c