add API to modify a node in-place, and option to preserve the location when cloning a node
Summary:
A very common usecase is to do
```
context.replaceNode(
  node,
  context.shallowCloneNodeWithOverrides(node, { ... }),
);
```

This diff adds an API, `modifyNodeInPlace` which provides a short-hand for this to make it easier:
```
context.modifyNodeInPlace(
  node,
  { ... },
);
```

Also adds a new flag to the internal detaching and cloning utils which allow us to preserve the location.
From what we've seen that most usecases of shallow cloning intend to reuse the node in-place, so they want this behaviour.
This is why I've set this to `true` for shallow clones.
Deep clones still blow away the location.

Reviewed By: pieterv

Differential Revision: D37774622

fbshipit-source-id: e680b47cc61d1e3eb9b427228ef7c3dc8cf622ce