Make WeakRef compliant
Summary:
If a WeakRef is constructed or dereferenced successfully, the target object must be kept alive until the end of the current JS job. We define the end of a job as a call to `drainMicrotasks` (consistent with V8 and JSC.)

To implement, we add a `PinnedHermesValue` in `Runtime`, `keptObjects_`. This is a `OrderedHashMap`. This way, we don't have to add any explicit checks for marking this data structure anywhere- it is already a GC-managed data structure.

Reviewed By: neildhar

Differential Revision: D36171397

fbshipit-source-id: ff8f9178344629404250299e89918871ad928409