Replace Date constructor
Summary: Convert `Date()` and `new Date()` to HostFunctions. This was a little trickier to pull off. The main reason is because they both return object/string ids which are not defined if you just return the value coming from `runtime_`. So the solution is to recreate the object/string in the tracing runtime with the value from the untraced `runtime_` once you run the nondeterministic function.

Reviewed By: kodafb

Differential Revision: D37229133

fbshipit-source-id: 0022e8c2b49d77c1ef5717b1e5026281d9a94bee