set parent pointers as part of the ESTree parse
Summary:
Do a final traversal of the AST at the end of the conversion to set the parent pointers.
In the ESLint world, these are set by ESLint just before the lint run. This isn't helpful for us because it means we can't use them within the hermes packages (like in hermes-eslint or hermes-transform).

Reviewed By: pieterv

Differential Revision: D38064433

fbshipit-source-id: 7e6d4c3556e3961797dd96d45aaa8fb01096b639