fix jsx above the react import not correctly marking react as used
Summary:
This was a bit bigger than I'd have liked.
Before the code would look in the upper scope for a variable of the required name.
This obviously doesn't work unless the variable has been seen at that point in the AST - i.e. it breaks if the import is after the usage.

Reviewed By: pieterv

Differential Revision: D38065365

fbshipit-source-id: a68a3434395bebc0d5982d509596969af3f6f1bd