Clear `next` when allocating list elements from freelist
Summary:
A bug in the allocation path resulted in bad list elements if they
were allocated off the free list.

The `next` pointer isn't cleared when we free the node (to avoid extra
work), but it must be cleared when we allocate elements at the end of a
list, to prevent accidental chaining onto previously existing list elements.

Reviewed By: tmikov

Differential Revision: D38095149

fbshipit-source-id: b0ac4ba7d491916ba062c1bc41b84afd2a39e8fc