Remove reserve in normalizeLocales
Summary:
Since the `length` is a JS controlled property, it is easy to pass in
an object with a large length, which can crash the process when
reserving space in the vector.

Given that this is not performance critical, we can remove the call to
`reserve` and let the loop over the elements handle ill-formed input.

Reviewed By: kodafb

Differential Revision: D38120294

fbshipit-source-id: 596fb31553cd7286222ccc02ffa8da121ac08f3f