Work around RTTI issues in CreateBigIntRecordFailures
Summary:
There seems to be a problem with duplicate typeinfo being generated for
`std::invalid_argument` on some compilers (e.g. on my laptop), which
means that the exception thrown by `std::stoull` cannot be caught with
the correct type, even though it is of the correct type.

Enabling RTTI for `SynthTraceParser` would also fix this issue, but we
currently can't use RTTI there because it consumes some classes that
are compiled without RTTI.

This is just a workaround to catch all exceptions so the test stops
failing, but we should put in a better fix soon.

Reviewed By: jpporto

Differential Revision: D38301033

fbshipit-source-id: 63d0b626619c9c13a10680214d1dd50144c54722