Use a dynamic threshold for incremental collections
Summary:
Dynamic thresholds were initially disabled for incremental collections
because of the possibility of it interacting with the drain rate in a
way that artificially ratchets up the drain rate and hurts pause times.
This diff makes the drain rate independent of the threshold, which gets
rid of the problem, and lets us use a dynamic threshold.

Reviewed By: jpporto

Differential Revision: D38299126

fbshipit-source-id: 22e6b2a099ae1b0b4cfc009b3360ec28982e39ae