Handle RTTI for types in public directory
Summary:
Export the typeinfo for types in the public directory from a single
place, so that anything consuming Hermes as a dynamic library will be
able to safely do things like `dynamic_cast`. This isn't particularly
important for the types currently defined there, but a later diff adds
a new exception type that does need correct handling for typeinfo, so
we might as well do it for the rest.

Reviewed By: jpporto

Differential Revision: D38317496

fbshipit-source-id: 3d540aed8977662f6b1b64d351bc2a04accdddd0