Ignore OOM exceptions in JSI layer
Summary:
Exception translation in JSI should ignore OOM exceptions, which are
unrecoverable, and should never hand back control to JS.

That is, an OOM exception must never be converted to a JS exception,
and should instead continue to be propagated as a C++ exception. This
is safe because OOM exceptions can only be thrown when Hermes is
compiled with exceptions enabled.

Reviewed By: jpporto

Differential Revision: D38295850

fbshipit-source-id: 96df9ebe168b1d25e7994ba2a4eae757418a56ec