Stop translating OOM exceptions
Summary:
Instead of converting OOM exceptions to `JSIException`s before
surfacing them, have them directly appear as `JSOutOfMemoryError`. This
makes it clear to the caller that an OOM has occurred and the VM may
not be reused.

This change is necessary to fix how OOMs thrown from native functions
are handled.

Reviewed By: jpporto

Differential Revision: D38295851

fbshipit-source-id: 769f06d212e381d6cca73a9d9a9d093f57a87efe