Remove enum_string from Apple DateTimeFormat
Summary:
We currently convert certain options for `DateTimeFormat` to an enum
before populating the entries. This adds additional complexity since it
requires an additional indirection. It also adds overhead, since we
need to perform unnecessary comparisons against values that are invalid
for a particular field, and then do an extra switch.

Reviewed By: jpporto

Differential Revision: D38415889

fbshipit-source-id: 0e7d835dd37a1da52cc77368c59e37a1a9afe2e7