Implement the CPU profiling for JS stack
Summary:
This diff stack adds support for loom integration for dApp, specifically capturing the stack of the JS function. The implementation hooks up the Hermes JS profiler to obtain the JS function stack and write into loom entries.
Here is the summary of 4 diffs in this stack:
1. Adds the ExternalApi which is the class that hooks up the Hermes JS profiler and the FBLoom.
2. Invoke the ExternalApi from Hermes via callback.
3. Create the FBLoomJSTraceProvider class, execute the callback when capturing loom.
4. Make the JS trace provider behind the mobile config.

In the test plan, we show a trace captured with this diff on Mac and Windows. On Windows, it doesn't do anything so the trace is empty. On Mac, we will see that the JS stack trace is labelled as "sampled js", which is different from the existing "sampled native".

Differential Revision: D37889426

fbshipit-source-id: 19eba2f4a374e7d043f2f5d5a0bbb0bd2d2b1e1a