Fix synth trace handling of Date constructor
Summary:
The current implementation of the `Date` constructor in synth traces
invokes `Date` by obtaining it from the global object. However, this
can break valid JS code, since the `Date` property may have been
changed by the user. We should instead be invoking the real `Date`
constructor that was saved.

See the added test for an example of how the current setup might
break.

To fix this, we can change the `nativeDateCtor` function to act like
`Date.now()`. The result of that call can then be used to
deterministically reconstruct the `Date` object.

Reviewed By: fbmal7

Differential Revision: D38897029

fbshipit-source-id: b6c73a90e1d1de33f4bcc07ede651d8b93b834f1