Add a simpler mechanism for tracing non-deterministic functions
Summary:
Add a new mechanism for tracing non-deterministic code that is primarily
in JavaScript. It uses a single HostFunction that acts as a trampoline,
taking the non-deterministic JS function as its argument, and calling it
without tracing.

This is much simpler than the current mechanism, and avoids the kind of
bugs that were identified in the previous diff.

However it does have the disadvantage of mixing references coming from
the tracing runtime with references coming from the base runtime. This
is safe, since it's in the implementation of TracingRuntime, where we
know the references are interchangeable, but it's still a drawback.

Convert `Math.random` to this new mechanism as a proof of concept.
Subsequent diffs will convert the rest.

Reviewed By: fbmal7

Differential Revision: D38969993

fbshipit-source-id: dd64d1687f6668cee3862cffcdfcd589ef0f763d