Re-work the TimeLimitMonitor implementation
Summary:
This change modifies the TimeLimitMonitor implementation so
* watching/unwatching Runtimes is fast
* auxiliary thread is kept alive as long as any Runtimes are registered with the TimeLimitMonitor
* the monitoring thread is only woken up when there's work to be done (i.e., it needs to check for timeouts).

The idea in this design is to have a global weak_ptr holding the //current// TimeLimitMonitor, if there is one, so all runtimes interested in time limit monitoring can reuse the same object. When the last Runtime unregisters itself from the TimeLimitMonitor (by reseting a shared_ptr) the timerLoop thread is cancelled and joined.

Reviewed By: neildhar

Differential Revision: D39230445

fbshipit-source-id: 0b0739cf8d4485e9d21ae9b595a9136e47a00e8d